requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.6
pydantic==2.9.2
scikit-learn>=1.3.0
pandas>=2.0.0
mysql-connector-python>=8.0.0
streamlit>=1.28.0
plotly>=5.17.0
openai>=1.0.0
requests>=2.31.0
